refactor(app): extract internal/app/httpx package (#32, step 1/5)
Moves the /readyz, /healthz, and /metrics-scrape HTTP wiring out of
cmd/topology-exporter/main.go and into a focused package alongside its
own tests.

Exports:
  - httpx.CycleStatus (was unexported cycleStatus in main)
  - httpx.NewReadyzHandler
  - httpx.NewHealthzHandler
  - httpx.InstrumentMetricsHandler

countingResponseWriter stays unexported in the new package; the four
panic-shim tests (issue #7) moved with it.

main.go: 1527 → 1427 lines. Four steps remain to land issue #32.

No behavior change. go build / go vet / go test -race / gofmt all clean.
